requirements.txt
# =====================================================================
#  KontentPyper -- Dependencies (curated, project-specific only)
# =====================================================================

# -- Framework & Server ------------------------------------------------
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
aiofiles>=23.0
python-multipart>=0.0.9

# -- Database (Neon PostgreSQL) ----------------------------------------
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# -- Auth & Security ---------------------------------------------------
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.1.0

# -- HTTP Client -------------------------------------------------------
httpx>=0.27.0

# -- Social Media OAuth & APIs ----------------------------------------
requests-oauthlib>=1.3.1
requests>=2.31.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2.0

# -- AI / LLM (LangChain + OpenAI) ------------------------------------
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
openai>=1.12.0

# -- Content Sources ---------------------------------------------------
praw>=7.7.0
feedparser>=6.0.0

# -- Media Processing --------------------------------------------------
Pillow>=10.0.0
moviepy>=1.0.3
edge-tts>=6.1.0

# -- Notifications -----------------------------------------------------
python-telegram-bot>=20.0

# -- Scheduling --------------------------------------------------------
apscheduler>=3.10.0

# -- Utilities ---------------------------------------------------------
python-dotenv>=1.0.0
